Extract icinga parsing logic into a class
